function/requirements.txt
azure-functions
sendgrid
psycopg2-binary
pytz
